+++ release/RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 # Evergreen Prompt
 
 # Changes
- 
+
+* Added "patch create" sub command.